# XL/bitrot: Migrate to new blake2b-simd SIMD optimized implementation.