… vendor/github.com/dchest/blake2b/README
… vendor/github.com/minio/blake2b-simd/LICENSE
… vendor/github.com/minio/blake2b-simd/README.md
… vendor/github.com/minio/blake2b-simd/blake2b.go
… vendor/github.com/minio/blake2b-simd/compress_amd64.go
… vendor/github.com/minio/blake2b-simd/compress_amd64.s
… vendor/github.com/minio/blake2b-simd/compress_generic.go
… vendor/github.com/minio/blake2b-simd/compress_noasm.go
… vendor/github.com/minio/blake2b-simd/cpuid.go
… vendor/github.com/minio/blake2b-simd/cpuid_386.s
… vendor/github.com/minio/blake2b-simd/cpuid_amd64.s
… vendor/vendor.json